Refactor directory.
