+++ case_studies/galaxy_clustering/data_generation/background_prior.py
@@ -1,11 +1,9 @@
 import os
-import pickle
 import random
 from pathlib import Path
 
 import numpy as np
 import pandas as pd
-from astropy import units
 from astropy.io import fits
 from scipy.stats import gennorm
 
@@ -26,26 +24,14 @@ class BackgroundPrior:
         self.reference_band = 1
         self.ra_cen = 50.64516228577292
         self.dec_cen = -40.228830895890404
-        self.mass_min = (10**14.5) * (units.solMass)
-        self.mass_max = (10**15.5) * (units.solMass)
         self.pixels_per_mpc = 80
         self.mean_sources = 0.004
-        self.mag_ex = 1.3
-        self.mag_max = 25
         self.G1_beta = 0.6
         self.G1_loc = 0
         self.G1_scale = 0.035
         self.G2_beta = 0.6
         self.G2_loc = 0
         self.G2_scale = 0.032
-        with open("gal_gmm_nmgy.pkl", "rb") as f:
-            self.gmm_gal = pickle.load(f)
-        self.redshift_alpha = 1.65
-        self.redshift_beta = 3.33
-        self.redshift0 = 0.9
-        self.gal_prob = 0.7
-        self.light_speed = 299792.46  # in km/s
-        self.sigma_DM15 = 1028.9  # in km/s
 
     def sample_n_sources(self):
         """Sample number of background sources.
