+++ case_studies/galaxy_clustering/data_generation/cluster_prior.py
